VDB-4395: skip test on mac

+++ test/scripts/test-prepare-sra-names.pl
@@ -1,3 +1,8 @@
+if ($^O =~ /darwin/) {
+    print "prepare-sra-names.sh does not work on Mac: test was scipped\n";
+    exit 0;
+}
+
 $acc = 'GSE118828';
 `rm -fr $acc` ; die if $?;
